fix: set a limit to the list of players
The request sent to get the list of all players sometimes fails on some
server. It happened for instance on a server counting about 50k players.
Lowering the number of players we want to get to 45k seems to fix this.
This patch changes the way to get list of players appending has many
parts of 45k players has needed to rebuild the whole list.

Fixes #1

Signed-off-by: Olmagzar <olmagzar@gmail.com>