fix: restore tournament participants count
For some reason, commit ac1ae4b introduced a condition on tournament
participants count. This resulted in not counting participants anymore
except when it comes to retrieving a single guild. Besides, the count in
this case was oftenly wrong if a previous database existed as it counted
the total active participants instead of the ones from the guild.

This patch restore general count of participant and select only
participant from the same guild when a guild is specified (either with
'--guild-name' or '--player-guild' option).

Signed-off-by: Olmagzar <olmagzar@gmail.com>